# Check `isbanded` in conversions