While converting from an `AbstractMatrix` to a `BandedMatrix`, we may use a faster path if the argument is banded and has a `BandedColumns{DenseColumnMajor}()` memory layout .
On master
```julia
julia> D = Diagonal(rand(1000));

julia> @btime convert(BandedMatrix, $D);
  4.521 μs (4 allocations: 8.11 KiB)

julia> @btime BandedMatrix{Float32}($D);
  4.588 μs (2 allocations: 4.11 KiB)

julia> @btime BandedMatrix($D);
  2.396 μs (2 allocations: 7.98 KiB)
```
This PR
```julia
julia> @btime convert(BandedMatrix, $D);
  759.925 ns (6 allocations: 8.20 KiB)

julia> @btime BandedMatrix{Float32}($D);
  564.160 ns (4 allocations: 4.20 KiB)

julia> @btime BandedMatrix($D);
  641.483 ns (4 allocations: 8.06 KiB)
```
With some refactoring, all of these could become equally performant.

TBH I'm a bit surprised that this is necessary at all, since I would have thought that banded broadcast already handles this. Presumably, much of the gain comes from linear indexing and vectorization.